// src/Services/models/PropertyTitleIEnumerableStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PropertyTitle } from './PropertyTitle';

export type PropertyTitleIEnumerableStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: Array<PropertyTitle> | null;
    statusCode?: string | null;
    errors?: any;
};