// src/Services/models/StringStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type StringStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: string | null;
    statusCode?: string | null;
    errors?: any;
};